BUG#16268281 : mysqlrplshow should show slave state
This patch allows mysqlrplshow to display additional information about the
slave state when verbosity is used (option --verbose or -v). More precisely,
now the utility also shows if the SQL thread is running or not (besides the IO
thread).

Existing MUT tests were corrected according to this change (for regression).