BUG#16664899 : add script return code checking to failover
This patch adds a feature to switchover and failover that 
permits the user to specify a threshold for checking the
return code of the external scripts. If the value returned
is GE the threshold, the operation will fail. 

Copious warning messages are provided. Default for the
threshold is to report return code for scripts but do not
fail.

Regression test cases added.

v2