BUG#14146713 : mysqldbimport cannot import multiple key col
This patch corrects a problem when indexes with multiple columns
are exported and imported. It corrects the defect whereby the columns
are repeated in the non-SQL formats and corrects the import code
to correctly add commas for multiple columns in the key definition
statement transformation.

A new test was added for regression.

v2