BUG#16499681 : mysqfrm does not read some files with special chars
This patch adds the capability to read file names with special codes
in the name. The format for these names is the output of the server
routines to embed special characters in file names. Examples include
'this@002ehas@002eperiods' and abc@002ekk@002dd@002ee@002df'. A
unit test was added to test the new code.

v2