BUG#16730148 :  mysqlauditadmin, mysqlauditgrep not found in mysqluc
This patch adds feedback to the user in the event that mysqluc cannot
detect a utility or read its options. It warns the user that the
utility will not appear in 'help utilities' and the utility is not
accessible from the console.

The error message, if present, is shown to the user or a tip is 
displayed if no error is detected.

New test added and test cases updated. Note that the test uses
Python 2.6 to generate the warning as documented in the bug
report.

v2