BUG#14762202 : fail to compare objects of different types with the same name
mysqldiff and mysqldbcompare failed to find differences when objects of
different types with the same name are found in the database. This patch fixes
this issues by distinguishing compared object by type and name.

In addition, minor usability issues were also fixed: correct the initial
message displayed by mysqldbcompare notifying the databases and servers being
compared (removing the reference to server2 when not specified), print the
correct message when errors are found, add '#' to the message indicating that
databases are consistent.

New MUT tests were added for regression and existing ones were corrected.