BUG#13864671 : mysqldbcopy fails to copy some views
This patch fixes mysqldbcopy that was failing to copy views, if there is no
space before the database name in fields selection in views. The database name
replacing mechanism for the create statement was simplified and added support
to spaces and parenthesis before the database name.

A MUT test was added for regression.