BUG#16745042 : output of mysqldbcompare has syntax errors
The mysqldbcompare utility was not generating a completely correct diff SQL
for routines that included statements ended with semi-colon (;) in the body,
leading to errors when executed with the mysql client. This patch fixes this
issue by surrounding the routine with a proper DELIMITER clause. In addition,
the DEFINER user@host and routine name in CREATE statements are now correctly
quoted with backticks.

A new MUT test case was added for regression.