BUG#16227817 : mysqldbimport fails with error as not all arguments converted
Running mysqldbimport with --format=csv and --import=data on a file that has
SET statements was raising an IndexError in is_quoted_with_backticks(). This
patch fixes this problem by adding the SET statements as definitions and not
as data, this doesn't apply to replication statements (started by --). If
there is no data in the file and the --import=data option is used, an error
message is now displayed.

A MUT test was added for regression.