BUG#16204629 : The mysqldbcompare utility can generate false hits.
The mysqldbcompare utility can generate false hits for missing or
changed data. This is because the algorithm uses a small key to
generate the checksum spans of the rows. The length of key has been
increased from 4 to 8, to allow more accurate hits. A new option was 
also added to allow the user to change this value. 

Regression tests was updated.
     
v3