BUG#16485689 : utilities fail when spaces are used in path
This patch corrects problems found during testing concerning the
change to serverclone.py. Specifically, the processing of the
mysqld options was incorrectly making them a string, the print
of the server command was not formatted correctly, and any 
options with extra spaces caused the Popen method to fail.

This patch also includes fixes for some regression tests.