BUG#16832161 : show variables don't work with long names in mysqluc
Using long variables names in mysqluc was raising a ValueError exception
while executing "show variables" command. This patch fixes this problem
by truncating the variables names when they are longer than max width
permitted. It was also fixed the support of passing multiple variables
to mysqluc.

A MUT test was added for regression.