BUG#16799466 : mysqldiskusage cannot find binlog usage
The mysqldiskusage utility was not displaying the binlog usage when used from
a remote host. This patch allows the binlog usage to be displayed using the
information from SHOW BINARY LOGS when the utility is executed from a remote
host. Now, the required privileges (to execute required queries) are also
checked before retrieving the logs usage and an appropriate error is issued.

Some additional minor fixes and improvements were also made to the
show_log_usage() function, namely: improve organization, follow coding style,
use of the default prefix when the relay log location (relay_log_basename) is
empty (for server > 5.6.2) to correctly display list of relay log files.

No MUT test was added to verify the main reported issue as it would require a
remote MySQL server. Nevertheless, some tests were added to verify the MySQL
access privileges to retrieve the logs usages.