BUG#16799550 : mysqlserverinfo cannot retrieve defaults from a remote server
The mysqlserverinfo utility issue a misleading error message when trying to
obtain information from a remote server. In particular, the defaults
information cannot be obtained from remote servers as it requires access to
local data. This patch fixes this issue, only retrieving the defaults data
when the utility is executed locally on the server and displaying an
appropriate warning message when a remote server is used.

In addition, some code improvements where performed to the serverinfo module,
namely: meet coding style, remove unused variables and code redundancies, and
more accurately detect if a started server is ready (new option 
--start-timeout added). Function is_alias() from the server module was also 
improved to more accurately detect aliases of 'localhost'.

MUT tests were added/changed to test the new option --start-timeout, but none
was added to verify the fix for the reported issue since it will require a
remote MySQL server.