BUG#16820207 : mysqlfailover wrongly fails with not enough privileges
The mysqlfailover utility was incorrectly recognizing the GRANT OPTION
privilege. This patches fixes this issue and checks if slaves have the
SELECT and RELOAD which are now also required to successfully execute 
failover due to recent improvements.

A new MUT test case was added for regression, checking all strictly required
privileges on slaves to successfully perform failover. In addition, the
failover test case was refactored to ensure that all temporary data is
correctly cleaned (kill spawned server). The failover_instances test was also
changed to remove the temporary directory of a killed server.