BUG#16920452 : mysqldiskusage crashes with databases with strange names
The mysqldiskusage utility was crashing when used against servers that
contained databases with strange names. This patch fixes the encoding of
databases to the correct directory name.

A MUT test was added for regression.

Note: the tools module was copied from 1.3.2 in order to reuse the encoding
methods used by the FRM reader, avoiding further conflicts when merging with
the 1.3 tree.