BUG#16953933 : reduce python version requirement to 2.6
This patch removes the requirement for Python 2.7 from the audit*
utilities.

Patch disables the utilities_console_warnings test as this test
contains test cases to check for warnings for checking invalid
Python version for the audit* utilities. Since the version is
now 2.6, these tests cases are not valid. We keep the test in
case we need a similar test case in the future.

Also included is a small change to MUT. The --force option was
removed from the disabled test check. Disabled tests must be
removed from the disabled file to be executed.