BUG#16748143 Improve MSI with shortcut to mysqluc
A shortcut opening mysqluc in a Command Prompt has been
added. It is installed in the Windows Start Menu. There is
also an option which allows the user the launch the
MySQL Utilities Console when the installation is finished.

More information was added to the exit dialog mentioning
that the utilities are available through the Command Prompt
or PowerShell session.