BUG#14348434 : Utilities do not permit --skip-grant-tables
This patch fixes a regression issue which lead to errant transaction when
executing failover utilities and undesired privileges changes. Now, SHOW
GRANTS is used instead of CREATE/DROP USER to determine if the option --skip-
grant-tables was used to start the server.

In addition, the result file of show_rpl test was fixed and some code 
cleanup and improvements were made to test file show_rpl_skip_grants.