BUG#16991428 : MUT replication tests spawn too many servers
When running the MUT suite a lot of MySQL server instances are created
making it impossible or extremely slow to run the entire suite in PCs or
VMs with low memory. This patch fixes that by killing some of the spawned
servers, i.e. servers that are only used in a single test and servers that
won't be used by later tests.

CAB v2