BUG#16900862 : mysqlindexcheck not finding all redundancies
The mysqlindexcheck utility was failing to identify clustered indexes
redundancies for secondary keys, i.e. secondary indexes that include the
clustered index (primary key). This patch allows the utility to identify this
type of redundancies and generate the appropriate ALTER TABLE statement (to
remove the clustered index columns from the index) when the --show-drops
option is used. Note that this kind of index redundancy is engine specific
and only applies to InnoDB storage engines.

MUT tests were added for regression.