BUG#16003529 : The test import_rpl runs inconsistently on windows
This patch fixes the timing issues present in the import_rpl and 
copy_db_rpl tests. Instead of a simple time.sleep to wait for the 
slaves to catch up, these tests now make use of the wait_for_slave
method from the Slave class. Also, a factory method was added to 
the Server class, which allows the creation of new servers
from existing server instances.