BUG#12889758 : use db pattern for --exclude in mysqldbcopy and mysqldbexport
The option --exclude already allows the use of SQL LIKE patterns since release
1.0.4. of mysql-utilities. By default SQL LIKE patterns are assumed, only
switching to SQL REGEXP patterns if the option --regexp is used. Nevertheless,
MUT tests were missing.

This patch add new MUT test for mysqldbcopy and mysqldbexport to verify the
use of SQL LIKE patterns for the option --exclude. In addition, it also
improves the use of specific object names for the --exclude option, now
supporting names quoted with backticks.