BUG#13103450 : mysqldbimport fails to import sakila database
The mysqldbimport utility was failing to import the Sakila sample database.
This patch fixes delimiters and multi-line queries handling, now allowing the
Sakila sample database to be successfully imported. In addition, the
identification of the database name in the import file have been improved,
privileges check have been optimized in order to be performed only once, and
trailing white spaces and delimiter are now removed from all queries to be
processed by the connector (since there are not required).

A new MUT test was added for regression and a small adjustment was made to
import_errors (remove trailing delimiter).