BUG#13773197 : mysqlserverclone complains it can't find mysqld
This patch fixes the error generated by mysqlserverclone when using a
remote host in the --server option. It checks if the host provided is
an alias of localhost and if it is not, shows a message saying that
mysqlserverclone does not work with remote hosts.

The is_alias() method from the server module was also improved to more
accurately detect aliases of 'localhost'. Now it handles local domain
names ending with '.lan'.

No MUT tests were added/changed because it involves remote hosts.