BUG#16918106 : let mysqlfailover run as daemon
This patch adds support for running mysqlfailover as a POSIX daemon. It's
possible to control the daemon by issue commands in the --daemon option, with
start, stop and restart. Running the failover daemon in foreground is also
possible by using 'nodetach' in the --daemon option.

This patch also ensures the registration of the failover instance on the
master is transferred to the new master on failover. It also ensures the
registration is removed from any slaves.

A MUT test was added for regression and existing ones were corrected
according to this change.