BUG#17019115 : mysqluc search "error" string instead of check return code
Upon starting the Utilities Console,  if there is an error parsing the
help messages from the other utilities, a warning message is printed to 
warn the user. However, if the string "ERROR"is found in one of the help 
message's text tried to parse, the same warning is also being 
printed and it shouldn't. This patch fixes that incorrect behavior.

In addition, On executing an utility command, Now all the error codes 
and output from stderr are saved to a list of errors which can be 
accessed through the new "show errors" command. Furthermore, it is also 
possible to check what was the last error that occurred using the 
"show last error" command and to erase the list of errors using the 
"clear errors" command.

v3