BUG#13103450 : mysqldbimport fails to import sakila database
The import_sakila test was failing because a different path was printed for
the imported file in windows and non-windows systems (forward slashes vs
backward slashes). This patch add a mask to print a fixed path independently
of the Operating System where the test is executed.

In addition, a few mask were added to fix an non-determinism issue with the
rpl_admin_gtid test.