BUG#16918106 : let mysqlfailover run as daemon
Errant transactions were introduced by the mysqlfailover utility because
unregistering the instance on all slaves was performed without disabling the
binary logging. This patch fixes this regression issue.

In addition, this patch also corrects the unregister of the failover instance
on the master when the failover utility is stopped or an error occurs.