Release-1.3.4 Testing
A regression was found in the serverclone.py module where the addition
of the --user option was overwritten. However, once replaced there
were additional problems found on Posix systems WRT running mysqlfrm
as root and requesting another user to run the spawned server under.

This patch corrects the problems by allowing the bootstrap to run
as root to ensure all files are copied, changes the owner and group
of the temporary datadir, then executes the server.

The temporary datadir is removed on completion.

v2