BUG#17217461 : mysqluserclone throws exception when --source omitted
If the user omits --source, the code attempts to use the default
but if the default credentials are not valid or the server does not
exist on port 3306, the code will thrown an exception.

Since there is no reason to execute the query without --source, this 
patch makes the --source option required for using mysqluserclone. 

Regression test added.

v2