BUG#13572964 : mysqlprocgrep missing kill process by id
The mysqlprocgrep utility did not provide any option to search and kill
processes by Id. This patch adds the --match-id option to search processes by
Id, allowing them to be killed by Id.

No MUT test was added for this specific option because it is not possible to
anticipate the Id assigned to a process.