BUG#16226348 : test server_info_errors does not execute correctly on windows
The server_info_errors test was unable to run successfully on windows and was 
therefore disabled. This happened because the error message returned
in cases where one cannot connect to a server is slightly different 
on windows than it is on Linux and it was not included in 
the existing code. This patch fixes that and enables the test on windows.