BUG#17474810 : constraint error copying the employees with mysqldbcopy
The mysqldbcopy utility was raising an error and exiting when cloning the
employees database as a result of failing to add a foreign key constraint.

This happened because the code was not correctly reproducing fkey constraints
for multiple constraints pointing to different tables. This patch fixes this
bug. 

The database used to test this utility was also improved in order to test
the several constraints/several tables issue that was causing this bug.