BUG#17510350 : mysqldbcompare fails with no error message
The issue described in this bug pertains mysqluc and not
mysqldbcompare.


The response given by mysqluc was confusing when an error code is returned
from the execution of a given utility. It appeared as if the
error message printed by mysqluc was generated by the utility itself
and not by mysqluc.

This patch fixes this issue by improving the error messages printed by
mysqluc (making them clearer) and adding a blank line to separate the
utility's output from the one of mysqluc. Furthermore, with this patch,
error messages are only saved when the utility streams an error
message to the standard error. Test cases were updated.