BUG#17393523 : Utilities can not be upgraded using RPM distribution pkg
When MySQL Utilities is installed using the RPM package, it was not 
possible to upgrade with a newer version or change from a license type 
with another RPM package without manually remove the first one. With this
patch, is now possible to upgrade version and so change from GPL license
to Commercial license without manually removing the previously installed
RPM package.

RPM spec now contains the Conflicts and the Obsoletes tags to tell that only
one package (GPL or Commercial) can be installed at a time, and one 
Obsoletes the other so the installed will be removed previews installing 
the other.