BUG#17548335 : clone_db test to copy databases with weird names not executed
The code of one of the clone_db tests was not reached and the test to clone
databases with weird names was not executed by MUT. This patch fixes this
issue by removing an incorrect return call. In order for the test to pass,
some corrections were also made to the database module (i.e., to quote
database objects with backticks).

In addition, an improvement was also made to issue a warning when the drop of
existing objects fails if the verbose mode is used. A MUT test was added to
test this enhancement.