BUG#17393742 : Debian pkgs can't use C/py if License types are different
MySQL Utilities Debian distribution packages could not able to install,
if the installed Connector Python had a different license type, i.e, 
both packages had to be using the same license. This patch makes it 
possible to install commercial version of Connector Python with GPL 
version of MySQL Utilities or vice-versa.      

The Debian Control files now contain Conflicts and Replaces tags to
indicate that the content from one package (GPL or Commercial) cannot
be installed along with the content from the other. In the past the
installation of a second package was not possible at all, and the
installation finished with errors of conflicting files.

In addition the copyright files that contained duplicate info from
README.txt and README_com.txt files were removed and are now created
upon the package building process.

v2