BUG#17622298 : mysqldbcopy and mysqldbexport copy and export routines after views
The mysqldbcopy and mysqldbexport utilities were creating views before
user-defined functions or procedures.
As result, the copy or export would fail if at least one of the views was
using user-defined functions or procedures.

This patch fixes this issue by switching the order in which objects are
created and exported. Now functions and procedures are processed before views.
In addition, a workaround was added to fix an issue regarding 5.1 servers and
the use of SHOW CREATE statements.

The util_test test database was updated with a new view which depends on a
function and the existing test cases were updated to reflect these changes.