BUG#17722274 : Error when copying db with blob fields
If a user attempts to use mysqldbcopy to copy a database with one or
more blob fields, the utility will fail with "TypeError: not all 
arguments converted during string formatting".

This is because there is a logic error in table.py in the 
_build_blob_update() method. It references an array when no array
exists.

This patch corrects the defect by referencing the string returned by
_build_blog_update() and adds a test case for regression as well as
updates existing test results that use the copy_db test.

v2