BUG#17555531 : Fabric Windows Installation fails in trying to use /etc
The MySQL Fabric configuration file was being set incorrectly to an 
unaccesible location. When installing from source, the configuration 
file is is configurated during the install time, and was using '/etc' 
for all platforms including Windows, now has been changed to 'etc'. 
In the case of the MSI installation the configuration file is 
configurated during the build of the MSI file, and was used a different 
location that the spected by the mysqlfabric script.

In addition an issue was corrected when doctrine version increments.