BUG#17066910 : a killed connection can cause failover
When the mysqlfailover connection to the master is killed, occasionally
it causes an automatic failover. With this patch, the utility will try
to reconnect to the master and after 3 attempts it will assume that the
master is down. Also the connection to the master is not maintained
between the intervals of reporting.