BUG#13417229 : mysqldbexport/mysqldbimport doesn't work with non-latin data
This patch adds the --character-set option in some utilities that
allows the user to define the client character set used for the MySQL
connection. If this option is not provided then the character set
retrieved from the server variable 'character_set_client' is used.

Existing MUT tests were corrected according to this change.