BUG#17633465 : mysqldbexport requires the replicate user on master
The replication user must indeed be created on the master in order for the
slaves to connect to it (not a bug). Nevertheless, the help description for
the --rpl option is not accurate and it can be misleading since no
destination/slave server can be specified for mysqldbexport. This patch
improves the help  description of the --rpl option in order to avoid
misinterpretations.

Existing MUT tests were updated for regression.