BUG#16386941 : Parsing errors to identify treatment instead of use errno
The mysqlserverinfo utility was parsing errors messages raised during 
server connection to suggest actions that the user can take about these 
errors (such as the use of the start option when the server is offline).
Now with this patch error numbers will be used for same purpose. In 
addition, the Exception from parse connection was change from UtilError 
to FormatError, to help to identify usage errors more appropriately.

Current general behavior was unchange, so mut tests were not updated.

v2