BUG#13563798 : mysqldbimport error when importing without --drop-first
When importing a valid import file in SQL format that includes a
CREATE DATABASE statement and the database already exists, mysqldbimport
does not issue a proper error message. This patch adds a proper error
message, informing the user to use the --drop-first to drop the database
before importing.
Also, the mysqldbcopy --force option was renamed to --drop-first.

A MUT test was added for regression and existing ones were corrected
according to this change.