BUG#17655980 : mut tests fail with connector/python 1.1
Some MUT tests were failing with C/Python version 1.1 due to changes
in the format of the error messages returned by the connector.

This patch fixes those tests, adding new masks or modifying existing ones
in order to make the tests work with both the GA and 1.1 versions of
the connector.