BUG#18117930 : mysqlserverclone fails to spawn server if datadir path is long
On Unix systems, the mysqlserverclone utility failed to spawn a server if the
path passed to the --new-data option was too long.
This happened because the socket was being stored inside the folder specified
by the --new-data option hence the path to the socket was too long.

This patch fixes this issue by creating a temporary folder and using it to
store the socket when the path would be too long. Also, a limit has been
applied to the length of the path passed to the --new-data option in order
to prevent issues with the cloning operation.

The use of sockets has also been removed from MUT and the verbose behavior was
changed. Now -vv provides the same debug  output as -vvv but without the
verbose output from launching and killing the MySQL servers.

Furthermore, the _remove_slave method from the topology module was fixed, it
now correctly identifies the same server even if under a different alias.

Existing tests were updated to reflect the changes introduced by this patch.