BUG#18041743 : foreign key constraint fails for mysqldbcopy with multiprocessing
When cloning a database with foreign key dependencies on the same server,
using the multiprocessing option (with a value greater than one), the
mysqldbcopy utility issues an error (foreign key constraint fails).

This patch fixes that issue by turning off the foreign key checks on the
server before cloning the tables.

The clone_db_fkeys test was updated to correcly exercise this issue.