BUG#13572814 : mysqldiskusage does not detect local vs remote server
This patch fixes mysqldiskusage that was using the local datadir
when using a remote server that has the same datadir.

Was added a proper local vs remote detection to the utility and
a new warning message is now shown if the user is connecting to
a remote server.

No MUT test was added, since a remote server connection is needed.