BUG#13492195 : dbcompare does not compare tables without primary keys.
The mysqldbcompare utility was not comparing tables that do not have a 
primary key. This patch adds a feature that permits users to use 
non-unique indexes. Only those unique indexes whose columns do not allow 
null values are permitted.

To use this feature, use the new option "--use-indexes" to specify
the unique index to be used to for the Table comparative. If the user
specifies an index that does not exist, an error will be shown. If the 
index specified allows null values, a warning will be shown however in 
this case the  next unique index will be used if one exists else the 
utility will stop with an error.

Regression tests was added to test the new option.
v2b