BUG#71498 : destination access check to strict for mysqldbcopy
The mysqldbcopy utility was always requiring a user SUPER privilege (even
to copy empty databases), instead of requiring the minimal privileges
allowed to perform the copy.

This patch fixes this issue by retrieving all types of objects from the
source database and testing for each one the privileges needed on the
destination database.

Furthermore, this patch introduces a TRIGGER privilege test for the
source database. This fixes the omission of triggers when copying a
database using a user that does not have the TRIGGER privilege on
the source database.

MUT tests were added for regression and existing ones were corrected
according to this change.