BUG#14725008: Reduced printed diff control lines for diff and dbcompare
The utilities mysqldiff and mysqldbcompare were printing all the control
lines for a printed difference; even if those lines were not really 
necessary as in the case of the grid format and unified type, which is 
the most notorious case. This patch adds a new option "--compact" to 
reduce the output from a diff.

Regression tests was added to test the new option.

v2. fixes broken diff_parameters test.