BUG#71654 : loginpath handling it not correct
The login-path handling from the utilities wasn't
generating helpful error messages if some of the required
options were missing.

This patch fixes this issue by providing an error message
which states which required options are missing.

This patch also removes the default values
for login-path options, forcing the user to be explicit about them. 

The mylogin test was updated to reflect these changes and two
new tests were added for regression testing.