BUG#17890455 : function gtid_enabled() does not check the full topology
The function gtid_enabled() from the mysql.utilities.common.topology.Topology
class was not correctly checking if GTID was enabled on the full topology. In
particular, the function was immediately returning the result for the first
checked server, independently if GTID is enable or not, without checking the
remains ones. This patch fixes the referred method, now correctly checking all
servers in the replication topology and only returning True if all servers
have GTID enabled (skipping not available/alive servers).

Two new MUT tests were added to validate this fix. In particular, the tests
were added to the rpl_admin_warnings tests, since the 'elect' operation of
the mysqlrpladmin utility is expected to issue a warning and exit when GTID
is not enabled on all servers (based on the result of the fixed function).