BUG#18275566 : method to find errant transaction fails with 5.6.9 servers
The function find_errant_transactions() from the Topology class (in module
mysql.utilities.common.topology) was not handling the server UUID in a
case-insensitive way. In the particular case of MySQL 5.6.9 servers, the
server UUID appears with different cases for the server_uuid variable (lower)
and in the GTID_EXECUTED set (upper), as a consequence the function was not
correctly ignoring the master transaction. This patch fixes this issue by
handling server UUIDs in a case-intensive way, more precisely when checking if
GTIDs belong to the master.

No new test is required because the 'rpl_admin_failover_error' and
'rpl_admin_failover_missing' MUT tests already test this issue and will fail
without the fix when using 5.6.9 servers.