BUG#17694651 : mysqldbcopy and mysqldbexport fail with views dependencies
The mysqldbcopy and mysqldbexport utilities fail copying/cloning or
exporting and then importing databases when there are dependencies
between views.

This patch detects the dependencies between the views and sort them in
the appropriate order before copying or exporting.

MUT tests were added for regression and existing ones were corrected
according to this change.