BUG#17878705 : mysqlfailover actual required privileges not match manual.
The privilege check for the mysqlfailover utility was performed after the
instance registration process, leading to query errors when some privileges
where not available (e.g. DROP and CREATE) instead of a permission error.
This patch fixes this issue by performing the privilege checks before the
instance registration operation. In addition, the check for the INSERT
privileges was added for all slaves, as it might be required for any of them
upon becoming the new master, in order to successfully perform the failover
instance registration on the new master.

In addition, the unused 'interval' parameter was removed from the
run_auto_failover() function in the RplCommands class
(mysql.utilities.command.rpl_admin module), and accordingly removed from all
function invocations in the code.

The 'failover_privileges' MUT tests were updated according to the performed
fix and also improved in order to test the privileges on the user for the 
master (not only the slaves). Additional variable renaming improvements were
performed in this test for coding style and code readability reasons:
    - the '_FAILOVER_LOG' variable was renamed to 'FAILOVER_LOG' (leading
	  underscore removed) in order to meet PEP8 naming conventions, since
	  it is a "module public" global variable;
    - the previous user name for slaves 'repl' was renamed to 's_user' to
      match the name convention used for the new master user 'm_user' (same
      for the password) in order to improve code readability;