BUG#17647498 : Unable to install Utilities from commercial zip package
During the setup procedure, when the Utilities scripts are built, the
python interpreter path was set to the path of the interpreter on the
PB2 build machine. This patch fixes this incorrect behavior, using
the path to the python interpreter of the host machine running the
script.
In addition, the distribution command modules were moved from the
"support" package to "support/distribution/commands/"  sub-package in
order to keep them separated from the other modules on the "support"
package.

v2. 
This version of the patch also includes changes to the commercial 
packages build mechanism to support the Fabric and Doctrine projects.