WL#6291 : replication synchronization checker
This patch fix a regression issue with the 'rpl_sync_errors' MUT test that
was failing when running the full MUT tests suite and by coincidence a
temporary server was still alive on localhost port 3333.

The tests to exercise invalid connection errors on the master and slave were
moved to ensure that a clean base command is used for them. The port used for
the invalid master connection test was changed to a value outside the range
of valid IPV4 ports, thus preventing it from being used by any temporary
servers spawned by MUT.