WL#7204 : Multi-source Replication
This worklog implements a multi-source replication that is often
regarded as a mean to aggregate and consolidate different streams of
data into one single server instance.

Since MySQL 5.6 does not support multi-source replication natively,
a round-robin scheduling was used to setup replication among all
masters and slave.

The support for running multi-source as a daemon was added for
POSIX systems.

MUT tests were added for regression.