BUG#18109179 : Unable to install fabric with --user option
The Utilities Setup was not able to install fabric configuration file
to users home directory when the '--user' install option is used, due 
a conflict path defined on Setup module used for the package mechanisms.
This patch fixes this problem by identifying when '--user' or '--home'
options are passed, to avoid overwrite the home-dir path with the path 
used for build packages.

In addition, the commercial MSI distribution mechanism has been modified
to include fabric and PHP doctrine if they are in the package path prior
to start the package build.

v3