BUG#18339351 : ports returned by get_next_port in mutlib can be in use
The get_next_port() method in mutlib was not testing if the port
number that it returned was already in use, leading to test failures
when a server was spawned for a port already in use


This patch fixes this issue for get_next_port() by testing if the port
to return is available, successively testing the next port until one
that is free is found.

Some existing tests were also corrected in order to make proper use of these changes.