Release-1.4.2 Testing
The MSI distribution package was missing the new utilities because 
the XML package descriptor needs to be updated after a new utility is 
created. This patch automates the process by adding the missing 
attributes for each of the utilities found in the scripts folder to 
the XML package descriptor file, at the time the build command is run 
for the GPL and commercial MSI distribution packages.

v1