BUG#18388846 : mysqlrplsync utility crashes when used with exclude option
The mysqlrplsync utility was crashing when the --exclude option was used and
at least one database was available to be checked. This issue was due to an
incomplete check when determining the tables to exclude from the slaves.

This patch fixes this issue by verifying if the checked database is included
in the dictionary of tables to exclude before checking if there is actually
any table to exclude. For reliability and safety purposes, the same
improvement was made to the corresponding check for tables to include.

MUT tests were updated to verify this issue.