BUG#18381748 : validate --exclude option with no value for mysqlrplsync
The mysqlrplsync utility was allowing the use of the --exclude option with an
empty value, not notifying the user about this situation.

This patch fixes this issue by issuing an error if the --exclude option is
used with an empty value. This option must be used with a valid value or not
used at all.

A new MUT test was added for regression.