BUG#18314951 : mysqlauditgrep does not support new audit log format
This patch fix a regression with the mysqlauditgrep utility, breaking the
compatibility with Python 2.6 due to a change in the latest patch. The dict
comprehension syntax using '{}' is not supported by Python 2.6 and earlier.
This patch replaces the dict comprehension syntax by the use of the dict
constructor.

No test new test added, existing MUT test for the audit log utilities now run
correctly with Python 2.6.