BUG#18376793:add missing __main__ attribute check for scripts directly executed
Most of the MySQL Utilities scripts that were directly executed at the
command line prompt (i.e., those in the scripts folder) did not
contain the if __name__ == '__main__' attribute check. This condition
check is a common practice to separate the code that can be imported
by other modules from the code that is accessed only when the script
is run as a standalone program.

This patch fixes this issue by adding this check to all the MySQL
Utilities scripts where it was missing.

Furthermore, it also fixes an issue where a database created by the
diskusage_parameters test wasn't being properly dropped making some tests
fail during the setup phase.

No MUT tests were added.