BUG#18281513 : mysqlindexcheck shows None instead of the drop statement
The mysqslindexcheck utility was considering as redundant a PRIMARY key 
when this key was compared to a unique index with more columns than the 
PRIMARY key, as a result; the drop statement was shown as "None;". 
 
This patch fixes this behavior in favor of the Primary key that should 
not be declared as a redundant in these cases, therefore the 
comparative mechanism was changed to mark the index key as a redundant 
key instead of the PRIMARY key, and now as a result the drop statement 
is for the unique index.

A test case was added to test this change.
v2